refactor: move input types to it's own file
